Optimize core enum sizes (from int to uint8_t)

+++ src/types.h
@@ -23,7 +23,7 @@
 #include <cstdint>
 
 // clang-format off
-enum Square : int {
+enum Square : uint8_t {
     a1, b1, c1, d1, e1, f1, g1, h1,
     a2, b2, c2, d2, e2, f2, g2, h2,
     a3, b3, c3, d3, e3, f3, g3, h3,
@@ -38,13 +38,13 @@ enum Square : int {
 
 constexpr int COLOR_OFFSET = 6;
 
-enum Color : int {
+enum Color : uint8_t {
     WHITE,
     BLACK,
     COLOR_NB
 };
 
-enum PieceType : int {
+enum PieceType : uint8_t {
     PAWN,
     KNIGHT,
     BISHOP,
@@ -55,7 +55,7 @@ enum PieceType : int {
     PIECE_TYPE_NB
 };
 
-enum Piece : int {
+enum Piece : uint8_t {
     WHITE_PAWN,
     WHITE_KNIGHT,
     WHITE_BISHOP,
